+++ Z80_Core/Memory/MemoryMap.cs
@@ -7,11 +7,11 @@ namespace Z80.Core
 {
     public class MemoryMap : IMemoryMap
     {
-        public const ushort PAGE_SIZE_IN_BYTES = 1024;
+        public const int PAGE_SIZE_IN_BYTES = 1024;
 
-        private Dictionary<ushort, IMemorySegment> _segments = new Dictionary<ushort, IMemorySegment>();
+        private Dictionary<int, IMemorySegment> _segments = new Dictionary<int, IMemorySegment>();
 
-        public uint SizeInBytes { get; private set; }
+        public int SizeInBytes { get; private set; }
 
         public IMemorySegment MemoryFor(ushort address)
         {
@@ -22,7 +22,7 @@ namespace Z80.Core
         public void Map(IMemorySegment entry, bool overwriteMappedPages = false)
         {
             ushort startAddress = entry.StartAddress;
-            uint size = entry.SizeInBytes;
+            int size = entry.SizeInBytes;
 
             if (startAddress % 1024 > 0)
             {
@@ -34,15 +34,15 @@ namespace Z80.Core
                 throw new MemoryMapException("Block would overflow the memory space which is " + SizeInBytes + " bytes.");
             }
 
-            ushort startPage = PageFromAddress(startAddress);
-            ushort endPage = (ushort)(startPage + (size / PAGE_SIZE_IN_BYTES) - 1);
+            int startPage = PageFromAddress(startAddress);
+            int endPage = startPage + (size / PAGE_SIZE_IN_BYTES) - 1;
 
             if (!overwriteMappedPages && _segments.Any(p => p.Key >= startPage && p.Key <= endPage))
             {
                 throw new MemoryMapException("Would overwrite existing mapped page/s. Specify overwriteMappedPages = true to enable masking the existing memory."); 
             }
 
-            for (ushort i = startPage; i <= endPage; i++)
+            for (int i = startPage; i <= endPage; i++)
             {
                 if (_segments.ContainsKey(i))
                 {
@@ -55,18 +55,18 @@ namespace Z80.Core
             }
         }
 
-        private ushort PageFromAddress(ushort address)
+        private int PageFromAddress(ushort address)
         {
-            ushort pageIndex = (ushort)Math.Floor((double)address / (double)(PAGE_SIZE_IN_BYTES));
+            int pageIndex = (int)Math.Floor((double)address / (double)(PAGE_SIZE_IN_BYTES));
             return pageIndex;
         }
 
-        private ushort AddressFromPage(ushort pageNumber)
+        private int AddressFromPage(ushort pageNumber)
         {
-            return (ushort)(pageNumber * PAGE_SIZE_IN_BYTES);
+            return pageNumber * PAGE_SIZE_IN_BYTES;
         }
 
-        public MemoryMap(uint sizeInBytes)
+        public MemoryMap(int sizeInBytes)
         {
             SizeInBytes = sizeInBytes;
         }
